requirements.txt
# =====================================
# الملاك المساعد - المتطلبات
# =====================================
# 
# متطلبات أساسية:
# - Python 3.8+
# - Windows 10/11 (للميزات الكاملة مثل F12 العالمي)
#
# تثبيت:
# pip install -r requirements.txt
# =====================================

# واجهة رسومية
PyQt6>=6.4.0

# التقاط الشاشة
mss>=9.0.0

# معالجة الصور
Pillow>=10.0.0

# طلبات HTTP لـ API
requests>=2.31.0

# تحميل متغيرات البيئة
python-dotenv>=1.0.0

# دعم Windows APIs (اختياري لكن مهم جداً)
# لميزة F12 العالمي + التعرف على النوافذ النشطة
pywin32>=305; sys_platform == 'win32'

# ✅ مكتبة keyboard للاختصارات العالمية (بديل موثوق)
keyboard>=0.13.5
